Fix cyclic dependency by adding relation_variant.
