+++ cxx/hirm.cc
@@ -3,173 +3,239 @@
 
 #include "hirm.hh"
 
-#include <cstdio>
-#include <ctime>
-#include <iostream>
-
-#include "cxxopts.hpp"
-#include "irm.hh"
-#include "util_io.hh"
-
-#define GET_ELAPSED(t) double(clock() - t) / CLOCKS_PER_SEC
-
-// TODO(emilyaf): Refactor as a function for readibility/maintainability.
-#define CHECK_TIMEOUT(timeout, t_begin)           \
-  if (timeout) {                                  \
-    auto elapsed = GET_ELAPSED(t_begin);          \
-    if (timeout < elapsed) {                      \
-      printf("timeout after %1.2fs \n", elapsed); \
-      break;                                      \
-    }                                             \
-  }
-
-// TODO(emilyaf): Refactor as a function for readibility/maintainability.
-#define REPORT_SCORE(var_verbose, var_t, var_t_total, var_model) \
-  if (var_verbose) {                                             \
-    auto t_delta = GET_ELAPSED(var_t);                           \
-    var_t_total += t_delta;                                      \
-    double x = var_model->logp_score();                          \
-    printf("%f %f\n", var_t_total, x);                           \
-    fflush(stdout);                                              \
-  }
-
-void inference_irm(std::mt19937* prng, IRM* irm, int iters, int timeout,
-                   bool verbose) {
-  clock_t t_begin = clock();
-  double t_total = 0;
-  for (int i = 0; i < iters; ++i) {
-    CHECK_TIMEOUT(timeout, t_begin);
-    single_step_irm_inference(prng, irm, t_total, verbose);
+HIRM::HIRM(const T_schema& schema, std::mt19937* prng) {
+  for (const auto& [name, relation] : schema) {
+    this->add_relation(prng, name, relation);
   }
 }
 
-void inference_hirm(std::mt19937* prng, HIRM* hirm, int iters, int timeout,
-                    bool verbose) {
-  clock_t t_begin = clock();
-  double t_total = 0;
-  for (int i = 0; i < iters; ++i) {
-    CHECK_TIMEOUT(timeout, t_begin);
-    // TRANSITION RELATIONS.
-    for (const auto& [r, rc] : hirm->relation_to_code) {
-      clock_t t = clock();
-      hirm->transition_cluster_assignment_relation(prng, r);
-      REPORT_SCORE(verbose, t, t_total, hirm);
-    }
-    // TRANSITION IRMs.
-    for (const auto& [t, irm] : hirm->irms) {
-      single_step_irm_inference(prng, irm, t_total, verbose);
-    }
+void HIRM::incorporate(std::mt19937* prng, const std::string& r,
+                       const T_items& items, const ObservationVariant& value) {
+  IRM* irm = relation_to_irm(r);
+  irm->incorporate(prng, r, items, value);
+}
+
+void HIRM::unincorporate(const std::string& r, const T_items& items) {
+  IRM* irm = relation_to_irm(r);
+  irm->unincorporate(r, items);
+}
+
+int HIRM::relation_to_table(const std::string& r) {
+  int rc = relation_to_code.at(r);
+  return crp.assignments.at(rc);
+}
+
+IRM* HIRM::relation_to_irm(const std::string& r) {
+  int rc = relation_to_code.at(r);
+  int table = crp.assignments.at(rc);
+  return irms.at(table);
+}
+
+RelationVariant HIRM::get_relation(const std::string& r) {
+  IRM* irm = relation_to_irm(r);
+  return irm->relations.at(r);
+}
+
+void HIRM::transition_cluster_assignments_all(std::mt19937* prng) {
+  for (const auto& [r, rc] : relation_to_code) {
+    transition_cluster_assignment_relation(prng, r);
   }
 }
 
-int main(int argc, char** argv) {
-  cxxopts::Options options("hirm",
-                           "Run a hierarchical infinite relational model.");
-  options.add_options()("help", "show help message")(
-      "mode", "options are {irm, hirm}",
-      cxxopts::value<std::string>()->default_value("hirm"))(
-      "seed", "random seed", cxxopts::value<int>()->default_value("10"))(
-      "iters", "number of inference iterations",
-      cxxopts::value<int>()->default_value("10"))(
-      "verbose", "report results to terminal",
-      cxxopts::value<bool>()->default_value("false"))(
-      "timeout", "number of seconds of inference",
-      cxxopts::value<int>()->default_value("0"))(
-      "load", "path to .[h]irm file with initial clusters",
-      cxxopts::value<std::string>()->default_value(""))(
-      "path", "base name of the .schema file", cxxopts::value<std::string>())(
-      "rest", "rest",
-      cxxopts::value<std::vector<std::string>>()->default_value({}));
-  options.parse_positional({"path", "rest"});
-  options.positional_help("<path>");
-
-  auto result = options.parse(argc, argv);
-  if (result.count("help")) {
-    std::cout << options.help() << std::endl;
-    return 0;
+void HIRM::transition_cluster_assignments(std::mt19937* prng,
+                                          const std::vector<std::string>& rs) {
+  for (const auto& r : rs) {
+    transition_cluster_assignment_relation(prng, r);
   }
-  if (result.count("path") == 0) {
-    std::cout << options.help() << std::endl;
-    return 1;
-  }
+}
 
-  std::string path_base = result["path"].as<std::string>();
-  int seed = result["seed"].as<int>();
-  int iters = result["iters"].as<int>();
-  int timeout = result["timeout"].as<int>();
-  bool verbose = result["verbose"].as<bool>();
-  std::string path_clusters = result["load"].as<std::string>();
-  std::string mode = result["mode"].as<std::string>();
-
-  if (mode != "hirm" && mode != "irm") {
-    std::cout << options.help() << std::endl;
-    std::cout << "unknown mode " << mode << std::endl;
-    return 1;
-  }
-
-  std::string path_obs = path_base + ".obs";
-  std::string path_schema = path_base + ".schema";
-  std::string path_save = path_base + "." + std::to_string(seed);
-
-  printf("setting seed to %d\n", seed);
-  std::mt19937 prng(seed);
-
-  std::cout << "loading schema from " << path_schema << std::endl;
-  auto schema = load_schema(path_schema);
-
-  std::cout << "loading observations from " << path_obs << std::endl;
-  auto observations = load_observations(path_obs, schema);
-  auto encoding = encode_observations(schema, observations);
-
-  if (mode == "irm") {
-    std::cout << "selected model is IRM" << std::endl;
+void HIRM::transition_cluster_assignment_relation(std::mt19937* prng,
+                                                  const std::string& r) {
+  int rc = relation_to_code.at(r);
+  int table_current = crp.assignments.at(rc);
+  RelationVariant relation = get_relation(r);
+  T_relation t_relation =
+      std::visit([](auto rel) { return rel->trel; }, relation);
+  auto crp_dist = crp.tables_weights_gibbs(table_current);
+  std::vector<int> tables;
+  std::vector<double> logps;
+  int* table_aux = nullptr;
+  IRM* irm_aux = nullptr;
+  // Compute probabilities of each table.
+  for (const auto& [table, n_customers] : crp_dist) {
     IRM* irm;
-    // Load
-    if (path_clusters.empty()) {
-      irm = new IRM(schema);
-      std::cout << "incorporating observations" << std::endl;
-      incorporate_observations(&prng, *irm, encoding, observations);
-    } else {
+    if (!irms.contains(table)) {
       irm = new IRM({});
-      std::cout << "loading clusters from " << path_clusters << std::endl;
-      from_txt(&prng, irm, path_schema, path_obs, path_clusters);
-    }
-    // Infer
-    std::cout << "inferring " << iters << " iters; timeout " << timeout
-              << std::endl;
-    inference_irm(&prng, irm, iters, timeout, verbose);
-    // Save
-    path_save += ".irm";
-    std::cout << "saving to " << path_save << std::endl;
-    to_txt(path_save, *irm, encoding);
-    // Free
-    free(irm);
-    return 0;
-  }
-
-  if (mode == "hirm") {
-    std::cout << "selected model is HIRM" << std::endl;
-    HIRM* hirm;
-    // Load
-    if (path_clusters.empty()) {
-      hirm = new HIRM(schema, &prng);
-      std::cout << "incorporating observations" << std::endl;
-      incorporate_observations(&prng, *hirm, encoding, observations);
+      assert(table_aux == nullptr);
+      assert(irm_aux == nullptr);
+      table_aux = (int*)malloc(sizeof(*table_aux));
+      *table_aux = table;
+      irm_aux = irm;
     } else {
-      hirm = new HIRM({}, &prng);
-      std::cout << "loading clusters from " << path_clusters << std::endl;
-      from_txt(&prng, hirm, path_schema, path_obs, path_clusters);
+      irm = irms.at(table);
     }
-    // Infer
-    std::cout << "inferring " << iters << " iters; timeout " << timeout
-              << std::endl;
-    inference_hirm(&prng, hirm, iters, timeout, verbose);
-    // Save
-    path_save += ".hirm";
-    std::cout << "saving to " << path_save << std::endl;
-    to_txt(path_save, *hirm, encoding);
-    // Free
-    free(hirm);
-    return 0;
+    if (table != table_current) {
+      irm->add_relation(r, t_relation);
+      std::visit(
+          [&](auto rel) {
+            for (const auto& [items, value] : rel->data) {
+              irm->incorporate(prng, r, items, value);
+            }
+          },
+          relation);
+    }
+    RelationVariant rel_r = irm->relations.at(r);
+    double lp_data =
+        std::visit([](auto rel) { return rel->logp_score(); }, rel_r);
+    double lp_crp = log(n_customers);
+    logps.push_back(lp_crp + lp_data);
+    tables.push_back(table);
+  }
+  // Sample new table.
+  int idx = log_choice(logps, prng);
+  T_item choice = tables[idx];
+
+  // Remove relation from all other tables.
+  for (const auto& [table, customers] : crp.tables) {
+    IRM* irm = irms.at(table);
+    if (table != choice) {
+      assert(irm->relations.count(r) == 1);
+      irm->remove_relation(r);
+    }
+    if (irm->relations.empty()) {
+      assert(crp.tables[table].size() == 1);
+      assert(table == table_current);
+      irms.erase(table);
+      delete irm;
+    }
+  }
+  // Add auxiliary table if necessary.
+  if ((table_aux != nullptr) && (choice == *table_aux)) {
+    assert(irm_aux != nullptr);
+    irms[choice] = irm_aux;
+  } else {
+    delete irm_aux;
+  }
+  free(table_aux);
+  // Update the CRP.
+  crp.unincorporate(rc);
+  crp.incorporate(rc, choice);
+  assert(irms.size() == crp.tables.size());
+  for (const auto& [table, irm] : irms) {
+    assert(crp.tables.contains(table));
   }
 }
+
+void HIRM::set_cluster_assignment_gibbs(
+    std::mt19937* prng, const std::string& r, int table) {
+  assert(irms.size() == crp.tables.size());
+  int rc = relation_to_code.at(r);
+  int table_current = crp.assignments.at(rc);
+  RelationVariant relation = get_relation(r);
+  auto f_obs = [&](auto rel) {
+    T_relation trel = rel->trel;
+    IRM* irm = relation_to_irm(r);
+    auto observations = rel->data;
+    // Remove from current IRM.
+    irm->remove_relation(r);
+    if (irm->relations.empty()) {
+      irms.erase(table_current);
+      delete irm;
+    }
+    // Add to target IRM.
+    if (!irms.contains(table)) {
+      irm = new IRM({});
+      irms[table] = irm;
+    }
+    irm = irms.at(table);
+    irm->add_relation(r, trel);
+    for (const auto& [items, value] : observations) {
+      irm->incorporate(prng, r, items, value);
+    }
+  };
+  std::visit(f_obs, relation);
+  // Update CRP.
+  crp.unincorporate(rc);
+  crp.incorporate(rc, table);
+  assert(irms.size() == crp.tables.size());
+  for (const auto& [table, irm] : irms) {
+    assert(crp.tables.contains(table));
+  }
+}
+
+void HIRM::add_relation(std::mt19937* prng, const std::string& name,
+                        const T_relation& rel) {
+  assert(!schema.contains(name));
+  schema[name] = rel;
+  int offset =
+      (code_to_relation.empty())
+      ? 0
+      : std::max_element(code_to_relation.begin(), code_to_relation.end())
+      ->first;
+  int rc = 1 + offset;
+  int table = crp.sample(prng);
+  crp.incorporate(rc, table);
+  if (irms.count(table) == 1) {
+    irms.at(table)->add_relation(name, rel);
+  } else {
+    irms[table] = new IRM({{name, rel}});
+  }
+  assert(!relation_to_code.contains(name));
+  assert(!code_to_relation.contains(rc));
+  relation_to_code[name] = rc;
+  code_to_relation[rc] = name;
+}
+
+void HIRM::remove_relation(const std::string& name) {
+  schema.erase(name);
+  int rc = relation_to_code.at(name);
+  int table = crp.assignments.at(rc);
+  bool singleton = crp.tables.at(table).size() == 1;
+  crp.unincorporate(rc);
+  irms.at(table)->remove_relation(name);
+  if (singleton) {
+    IRM* irm = irms.at(table);
+    assert(irm->relations.empty());
+    irms.erase(table);
+    delete irm;
+  }
+  relation_to_code.erase(name);
+  code_to_relation.erase(rc);
+}
+
+double HIRM::logp(
+    const std::vector<std::tuple<std::string, T_items, ObservationVariant>>&
+    observations) {
+  std::unordered_map<
+      int, std::vector<std::tuple<std::string, T_items, ObservationVariant>>>
+      obs_dict;
+  for (const auto& [r, items, value] : observations) {
+    int rc = relation_to_code.at(r);
+    int table = crp.assignments.at(rc);
+    if (!obs_dict.contains(table)) {
+      obs_dict[table] = {};
+    }
+    obs_dict.at(table).push_back({r, items, value});
+  }
+  double logp = 0.0;
+  for (const auto& [t, o] : obs_dict) {
+    logp += irms.at(t)->logp(o);
+  }
+  return logp;
+}
+
+double HIRM::logp_score() const {
+  double logp_score_crp = crp.logp_score();
+  double logp_score_irms = 0.0;
+  for (const auto& [table, irm] : irms) {
+    logp_score_irms += irm->logp_score();
+  }
+  return logp_score_crp + logp_score_irms;
+}
+
+HIRM::~HIRM() {
+  for (const auto& [table, irm] : irms) {
+    delete irm;
+  }
+}
+
