+++ cxx/tests/BUILD
@@ -2,7 +2,7 @@ cc_binary(
     name = "test_hirm_animals",
     srcs = ["test_hirm_animals.cc"],
     deps = [
-        "//:headers",
+        "//:hirm_lib",
         "//:util_distribution_variant",
         "//:util_io",
         "//distributions",
@@ -13,7 +13,6 @@ cc_binary(
     name = "test_irm_two_relations",
     srcs = ["test_irm_two_relations.cc"],
     deps = [
-        "//:headers",
         "//:irm",
         "//:util_distribution_variant",
         "//:util_io",
@@ -25,7 +24,8 @@ cc_binary(
     name = "test_misc",
     srcs = ["test_misc.cc"],
     deps = [
-        "//:headers",
+        "//:hirm_lib",
+        "//:irm",
         "//:util_distribution_variant",
         "//:util_io",
         "//distributions",
