+++ cxx/util_distribution_variant.hh
@@ -1,8 +1,9 @@
 // Copyright 2024
 // See LICENSE.txt
 
-// This file collects classes/functions that depend on the set of distribution
-// subclasses and should be updated when a new subclass is added.
+// Classes and functions for dealing with Distributions and their values in a
+// generic manner.  When a new subclass is added, this file and
+// relation_variant.{hh,cc} will need to be updated.
 
 #pragma once
 
@@ -11,6 +12,11 @@
 #include <variant>
 #include <vector>
 
+class BetaBernoulli;
+class Bigram;
+class DirichletCategorical;
+class Normal;
+
 enum class DistributionEnum { bernoulli, bigram, categorical, normal };
 
 struct DistributionSpec {
@@ -18,22 +24,11 @@ struct DistributionSpec {
   std::map<std::string, std::string> distribution_args = {};
 };
 
-class BetaBernoulli;
-class Bigram;
-class DirichletCategorical;
-class Normal;
-class Domain;
-template <typename DistributionType>
-class Relation;
-
 // Set of all distribution sample types.
 using ObservationVariant = std::variant<double, int, bool, std::string>;
 
 using DistributionVariant =
     std::variant<BetaBernoulli*, Bigram*, DirichletCategorical*, Normal*>;
-using RelationVariant =
-    std::variant<Relation<BetaBernoulli>*, Relation<Bigram>*,
-                 Relation<DirichletCategorical>*, Relation<Normal>*>;
 
 ObservationVariant observation_string_to_value(
     const std::string& value_str, const DistributionEnum& distribution);
@@ -41,7 +36,3 @@ ObservationVariant observation_string_to_value(
 DistributionSpec parse_distribution_spec(const std::string& dist_str);
 
 DistributionVariant cluster_prior_from_spec(const DistributionSpec& spec);
-
-RelationVariant relation_from_spec(const std::string& name,
-                                   const DistributionSpec& dist_spec,
-                                   std::vector<Domain*>& domains);
